+++ Gringotts.Api/Features/Statistics/Models/ReaderAnalytics.cs
@@ -1,4 +1,4 @@
-namespace Gringotts.Api.Shared.Database.Models.Readers
+namespace Gringotts.Api.Features.Statistics.Models
 {
     public record ReaderAnalytics
     {
