+++ Gringotts.Api/Features/User/Models/Role.cs
@@ -1,4 +1,4 @@
-namespace Gringotts.Api.Shared.Database.Models.Users
+namespace Gringotts.Api.Features.User.Models
 {
     public record Role
     {
