--- Gringotts.Api/Shared/Database/Models/Records/ActionType.cs
@@ -1,7 +1,0 @@
-namespace Gringotts.Api.Shared.Database.Models.Records
-{
-    public enum ActionType
-    {
-        Entry = 1, Exit = 2
-    }
-}
